issue-52 24.09.2023 Added TokenInsertion class Reworked TokenValidation class Made TokenAlreadyExists final
