+++ src/main/java/io/blamer/hub/exceptions/TokenAlreadyExists.java
@@ -1,6 +1,6 @@
 package io.blamer.hub.exceptions;
 
-public class TokenAlreadyExists extends RuntimeException{
+public final class TokenAlreadyExists extends RuntimeException{
     public TokenAlreadyExists(String message) {
         super(message);
     }
